Adds an example how we could log the entry/exit of HMC calls
As a library, we don't want to flood the logs of the library user
but want to offer the capability to grab the log records if needed.
This change adds a convenience function ``get_logger`` which adds
the NullHandler per default to prevent the flooding. It also adds
a decorator we can apply on methods which do calls to the HMC.
An added doc file explains how the user can grab the log records
this library emits.