[WIP] RFCv3 - zhmcsh: a command line tool for simple HMC
This is a request for comments: I suggest to add a command line processor
to facilitate simple HMC management tasks for people without Python
programming skills.

The current prototype is fairly self-explanatory and can be easily extendend
due to the use of the argparse module.
It is written such that it can be invoked by either using the script name
followed by an optional object class and a subcommand, or or by a symlink
for a given object class and a subcommand, e.g.

zhmcsh cpc list  hmc.ibm.com
or
zhmccpc list hmc.ibm.com

Initially, zhmcsh can display the API version, list CPCs and LPARs (or
partitions) for a CPC, and display details about those. Additional
objects and commands are rather easy to add, requiring only a few
declarations for the argparse module and a few lines of handler code.

Additionally the script acts as an HMC 'shell', i.e. allowing to perform a
sequence of commands (hence the -sh ending, instead of the -cli that might
have been expected). The shell mode keeps track of credentials per HMC used,
and offers simple command line editing capabilities through the readline
module.

As I am not a Python hacker (which is why I have written the script),
I am looking forward to constructive criticism and improvement
suggestions.

Signed-off-by: Viktor Mihajlovski <mihajlov@linux.vnet.ibm.com>