fix the unit tests for the session object
The "session_id" method of the Session object needs to be a read-only
property, otherwise the unit-tests in "test_session" will fail.
An alternative would have been to change the calls in the unit tests
from "session.get_session" to "session.get_session()", but this would
not fit to the rest of the code.