Added send_os_command method to partition and lpar
send_os_message allows the client to send messages to the object's operating
system. The method takes in a string for the command and and optional boolean
to indicate a priority. The method returns nothing because the HMC API returns
a response with no content. A status code 409 is returned from the API if the
message interface is not available.

Signed-off-by: Andrew Brezovsky <abrezovsky@gmail.com>