Fixed unconditional addition of ifl-processors in CLI.
This fixes issue #258, where a previous addition to the
development version of v0.13.0 has caused the
ifl-processors property to always be added to the Create
Partition request, if the corresponding option had not been
specified in the zhmc command line. Now, a default is only
added if neither CPs nor IFLs are specified on the command line.

Also, in order to simplify the logic, if the IFL or CP option
is specified with 0, that is no longer handled by removing the
option from the properties sent to the HMC. Instead, the HMC
is now going to reject that. This is more in line with having
easy to understand logic on the client side.

Signed-off-by: Andreas Maier <maiera@de.ibm.com>