Fixed find(name) on element resources.
The find(name) method on element-kind of resources raised NotFound
for existing resources. This was caused by the fact that the
list() method for element-kind of resources returns Python resource
objects that do not have a "name" property, so the operation that
is supposed to populate the name-URI-cache does not add anything.
Because of the optimization in find(name) to always use the
name-URI-cache, it relies on that cache being populated.
See issue #264.

This change fixes that by introducing knowledge into the manager
class about whether the list() method for a resource class
populates the "name" property. If that is not the case, the
refresh() method of the name-to-URI cache sets the 'full_properties'
flag when calling the list() method. For the element-kind of
resource classes, this drives Get {Resource} Properties HMC operations
to populate the cache, but afterwards, the cache is utilized.
This has been done for all element-kind of resource classes, i.e.
NICs, HBAs, VFs, Ports.

In addition, the unit tests for element-kind of resource classes
have been extended with test cases for filtering with name and with
element-id.

Signed-off-by: Andreas Maier <maiera@de.ibm.com>