CPI: Improvements for dealing with LPAR status 'exceptions'.
This change provides some improvement for the zhmc CLI when dealing
with the LPAR status "exceptions" when activating, deactivating or
loading the LPAR.

The main change is to add a command line option `--allow-status-exceptions`
for the `lpar activate/deactivate/load` commands. Setting this option causes
the LPAR status "exceptions" to be considered an additional valid end status.

Some other changes were needed as part of that:

* Allowed `None` as a value for the `load_parameter` argument of
  `Lpar.load(), and changed the default to be `None` (which does not change
  the behavior).

* Added actual status, desired statuses and status timeout as attributes to
  the `StatusTimeout` exception, for programmatic processing by callers.
  This is not yet exploited in the zhmc CLI, but it can already be
  exploited by users of the zhmcclient API.

Signed-off-by: Andreas Maier <maiera@de.ibm.com>