Fix ssc-dns-servers option handling in CLI
Details:
When we try to set the --ssc-dns-servers in the
zhmc partition create or update command we get
an HTTPError: 400,7 because the HMC expects
an array for this option. This change fills
the passed ssc-dns-servers separated by a comma
into an array. A maximum of two ssc-dns-servers
are accepted by the HMC.
Resolves #310

Signed-off-by: Juergen Leopold <leopoldj@de.ibm.com>