Add CLI support for mount and unmount an ISO
With the new CLI command 'zhmc partition mountiso'
an user provided ISO image can be uploaded.
The uploaded ISO image is the image that gets attached
to the partition. If an ISO is attached to a partition,
it can be shown by the partition property 'boot-iso-image-name'.
To boot from this ISO image the partition property
'boot-device' must be set to 'iso-image' explicitly
before starting the partition or you can add the '--boot'/'-b'
option to the 'zhmc partition mountiso' command.

With the new CLI command 'zhmc partition unmountiso'
the user unmounts the currently mounted ISO from
the partition. The partition property 'boot-iso-image-name'
is empty then.

Signed-off-by: Juergen Leopold <leopoldj@de.ibm.com>