Implementation of DPM storage group support for FCP
Note that this is only a subset of the entire DPM storage management
support: Missing is the entire FICON support, both for storage groups
and for the FICON-specific model of storage sites, storage subsystems,
etc.

Note that mock support and tests for the FCP storage group support is
only an initial subset for storage groups (but not for storage volumes
or virtual storage resources).

Details:

* New resource classes StorageGroup, StorageVolume, and
  VirtualStorageResource.

* New methods on resource class Cpc:
  `list_associated_storage_groups()`, `validate_lun_path()`.

* New methods on resource class Partition:
  `attach_storage_group()`, `detach_storage_group()`,
  `list_attached_storage_groups()`.

* New property `storage_groups` on resource class Console.

* Unit tests for storage groups.

* Mock support and mock unit tests for storage groups.

* Example script examples/list_storage_groups.py.

Signed-off-by: Andreas Maier <maiera@de.ibm.com>