Consolidated internal nic_/partition_object() methods to resource_object()
Details:

The NicManager and PartitionManager classes had an internal nic_object()
and partition_object() method, for generating a minimalistic Python resource
object of the respective resource type.

There is a need to add more such functions in the future, and because these
methods are fairly similar, this change generalizes these methods into
one BaseManager.resource_object() method.

The new generalized method adds the ability to get the object ID as input
instead of just the object URI. In addition, it adds the ability to get
additional resource properties (in addition to the minimalistic set of
resource properties that is generated automatically).

The test cases have been adjusted and extended accordingly.

Signed-off-by: Andreas Maier <maiera@de.ibm.com>